pyproject.toml
[project]
name = "cis-request-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docusign-esign>=5.4.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "jwt>=1.4.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
]
